SCC-2206 Add GET endpoint for single holding records
This adds a new endpoint to retrieve single holding records through a `GET` request to `/holdings/{id}`. This is being added to support functionality in the `discovery-store-poster` and potentially other applications that want an easy way to retrieve a single holding record.

This returns the holding object in the body of the response, not in an array as the other `GET` endpoint does