NOREF Convert column names to camelCase
This may be a bit unorthodox, but this converts the postgres column names to camelCase and updates this function to interact with this new schema. This is for three reasons:

- It aligns the Kinesis stream and API response for the holdings data to a single format
- It aligns the Holding stream with the Bib and Item streams and API responses
- It provides the fields in the format expected by the discovery store poster

To do so there are also some minor changes to the test suite and how queries are constructed when a `GET` request is received.

To deploy this we will have to make sure that the qa/production databases are updated to new schemas before deploying this function. (Ideally this would be done with an ActiveRecord migration, but that is overkill at the moment). So the procedure should be:

1) Temporarily disable the incoming Kinesis stream to the `HoldingPoster`
2) Manually update the table schema of the database
3) Deploy this new code to the function
4) Re-enable the Kinesis stream